Function for allocating handle returns zero by default
Related to #1 enhancement.
Add test to check whether HANDLE_Alloc() function returns zero handle by default. The function uses
the following components:
 - handle-to-memory look-up table
 - function for finding first empty LUT entry
 - memory allocation and entry fields initialization
The HANDLE functions have to be executed after HANDLE_Init() was called. If the init call is missing
the API functions will not be working properly.
The macro HANDLE_LUT_DEFAULT_SIZE defines the LUT default size and can be modified in the future.