Add function for deallocating handles
Related to #1 enhancement.
Add function HANDLE_Dealloc which deallocates a handle and enables it to
be reused. It searches for the handle entry using the FindLutEntry
helper function. The function has missing checking whether handle is
present in the LUT (this will be fixed in the following commits).
Add unit test to check if freed handle can be used second time